refactor : user_step 테이블 위치 변경 및 테이블명 수정
