+++ src/main/java/com/example/tily/user/User.java
@@ -2,7 +2,7 @@ package com.example.tily.user;
 
 import com.example.tily.BaseTimeEntity;
 import com.example.tily.roadmap.relation.UserRoadmap;
-import com.example.tily.step.UserStepRelation;
+import com.example.tily.step.relation.UserStep;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -26,7 +26,7 @@ public class User extends BaseTimeEntity {
     private List<UserRoadmap> userRoadmaps = new ArrayList<>();
 
     @OneToMany(mappedBy =  "user")
-    private List<UserStepRelation> userStepRelations = new ArrayList<>();
+    private List<UserStep> userSteps = new ArrayList<>();
 
     @Column(length = 50, nullable = false, unique = true)
     private String email;
